test: 테스트 중복 코드 추출
