+++ src/test/java/webserver/RequestDispatcherTest.java
@@ -1,48 +1,21 @@
 package webserver;
 
 import model.User;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import webserver.domain.QueryParameter;
 import webserver.domain.Request;
-import webserver.view.NetworkInput;
+import webserver.support.RequestHelper;
 
-import java.io.ByteArrayInputStream;
-import java.io.InputStream;
+import java.io.IOException;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-class RequestDispatcherTest {
-    // TODO 중복 제거
-    private final String requestInput =
-            "GET /index.html HTTP/1.1\n" +
-                    "Host: localhost:8080\n" +
-                    "Connection: keep-alive\n" +
-                    "Accept: */*";
-    private final String requestInputWithQueryString =
-            "GET /user/create" +
-                    "?userId=javajigi&password=password" +
-                    "&name=%EB%B0%95%EC%9E%AC%EC%84%B1&email=javajigi%40slipp.net HTTP/1.1\n" +
-                    "Host: localhost:8080\n" +
-                    "Connection: keep-alive\n" +
-                    "Accept: */*";
-    private InputStream inputStream;
-    private InputStream inputStreamWithQueryString;
-    private NetworkInput networkInput;
-    private NetworkInput networkInputWithQueryString;
-
-    @BeforeEach
-    void setUp() {
-        inputStream = new ByteArrayInputStream(requestInput.getBytes());
-        inputStreamWithQueryString = new ByteArrayInputStream(requestInputWithQueryString.getBytes());
-        networkInput = new NetworkInput(inputStream);
-        networkInputWithQueryString = new NetworkInput(inputStreamWithQueryString);
-    }
+class RequestDispatcherTest extends RequestHelper {
 
     @Test
     @DisplayName("url 분기처리를 제대로 하는지 확인")
-    void forward() {
+    void forward() throws IOException {
         final Request request = new Request(networkInputWithQueryString);
 
         final QueryParameter queries = request.getQueryParameter();
