+++ src/test/java/webserver/domain/RequestTest.java
@@ -1,43 +1,21 @@
 package webserver.domain;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import webserver.support.RequestHelper;
 import webserver.view.NetworkInput;
 
 import java.io.ByteArrayInputStream;
+import java.io.IOException;
 import java.io.InputStream;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-class RequestTest {
-    private final String requestInput =
-            "GET /index.html HTTP/1.1\n" +
-            "Host: localhost:8080\n" +
-            "Connection: keep-alive\n" +
-            "Accept: */*";
-    private final String requestInputWithQueryString =
-            "GET /user/create" +
-            "?userId=javajigi&name=pobi HTTP/1.1\n" +
-            "Host: localhost:8080\n" +
-            "Connection: keep-alive\n" +
-            "Accept: */*";
-    private InputStream inputStream;
-    private InputStream inputStreamWithQueryString;
-    private NetworkInput networkInput;
-    private NetworkInput networkInputWithQueryString;
-
-    @BeforeEach
-    void setUp() {
-        inputStream = new ByteArrayInputStream(requestInput.getBytes());
-        inputStreamWithQueryString = new ByteArrayInputStream(requestInputWithQueryString.getBytes());
-        networkInput = new NetworkInput(inputStream);
-        networkInputWithQueryString = new NetworkInput(inputStreamWithQueryString);
-    }
+class RequestTest extends RequestHelper {
 
     @Test
     @DisplayName("쿼리없는 객체 생성 테스트")
-    void create() {
+    void create() throws IOException {
         InputStream expectedInputStream = new ByteArrayInputStream(requestInput.getBytes());
         NetworkInput expectedNetworkInput = new NetworkInput(expectedInputStream);
 
@@ -52,7 +30,7 @@ class RequestTest {
 
     @Test
     @DisplayName("쿼리가 있는 객체 생성 테스트")
-    void createWithQueryString() {
+    void createWithQueryString() throws IOException {
         InputStream expectedInputStream = new ByteArrayInputStream(requestInputWithQueryString.getBytes());
         NetworkInput expectedNetworkInput = new NetworkInput(expectedInputStream);
